Factor 'generate_signed_url' into a free function in 'gcloud.credentials'.
Leave a wrapper method in 'gcloud.storage.connection.Connection'.

Fixes #57.
